+++ docs/changes.rst
@@ -13,6 +13,7 @@ Next version
   ``@override_settings``, to reconfigure the toolbar during tests.
 * Optimize rendering of SQL panel, saving about 30% of its run time.
 * New records in history panel will flash green.
+* Automatically update History panel on AJAX requests from client.
 
 3.2.4 (2021-12-15)
 ------------------